Added Animator3D and extracted common animation logic to Animator class
